requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb>=0.4.22
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.15
diskcache==5.6.3
durationpy==0.9
effdet==0.4.1
emoji==2.14.0
eval_type_backport==0.2.2
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub>=0.23.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.8
kubernetes==31.0.0
langchain>=0.2.0
langchain-community>=0.0.16
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langchain-huggingface>=0.0.2
langdetect==1.0.9
langsmith==0.2.7
layoutparser==0.3.4
libmagic==1.0
llama_cpp_python>=0.2.38
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.10.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy>=1.24.0
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.13
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pi_heif==0.21.0
pikepdf==9.5.0
pillow==11.0.0
portalocker==3.1.1
posthog==3.7.4
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic>=2.5.3
pydantic-settings==2.7.1
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.1
pypdf>=4.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic>=0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
pytz==2024.2
PyYAML>=6.0.1
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich>=13.7.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers>=2.5.1
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
tesseract==0.1.3
threadpoolctl==3.5.0
timm==1.0.12
tokenizers==0.21.0
torch>=2.0.0
torchvision==0.20.1
tqdm>=4.66.1
transformers>=4.38.0
typer>=0.9.0
typing-inspect==0.9.0
typing_extensions>=4.9.0
tzdata==2024.2
unstructured>=0.12.0
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
rank_bm25>=0.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.2
poppler-utils
pytesseract
Pillow>=10.1.0
qdrant-client>=1.7.0
